Hello, world! A minimum viable dynasm-based JIT
Add a stupid JIT pass based on the dynasm framework originally
developed for luajit.  It works by basically unrolling the bytecode
interpreter loop directly to x64 machine code with hand-allocated
registers.  Right now it only supports two instructions: LOAD_CONST
and RETURN_VALUE.

Eventually, I will want to add an intermediate pass that tries to keep
stack manipulation in registers, but I'm starting with the simplest,
stupidest thing possible.

The bulk of the JIT-related code is in `cjit.c`, with a bit of
dispatch code in `ceval.c`.  Right now, every time a stack frame is
evaluated, we check to see if JIT code exists for the corresponding
code object.  If not, we attempt JIT-compilation, falling back to the
interpreter if we encounter instructions we don't know how to JIT.

A copy of dynasm (both the preprocessor and the runtime library) is
bundled.  Right now, we still depend on a lua interpreter to run the
preprocessor.  I will probably need to rewrite the preprocessor in
Python.

Example of it working:

    >>> def myfun():
    ...     return 9
    ...
    >>> myfun()
    Linking and encoding
    Generated JITed code at 0x7f4b71697000, jumping
    Returned from JITed code
    9
    >>>